+++ cupy/random/_permutations.py
@@ -1,4 +1,4 @@
-from cupy.random import _generator
+from cupy.random import _random_state
 
 
 def shuffle(a):
@@ -10,7 +10,7 @@ def shuffle(a):
     .. seealso:: :meth:`numpy.random.shuffle`
 
     """
-    rs = _generator.get_random_state()
+    rs = _random_state.get_random_state()
     return rs.shuffle(a)
 
 
@@ -27,5 +27,5 @@ def permutation(a):
 
     .. seealso:: :meth:`numpy.random.permutation`
     """
-    rs = _generator.get_random_state()
+    rs = _random_state.get_random_state()
     return rs.permutation(a)
